Add OpenAPI validation and routing
Routing was supposed to be much more modular, but Slim does not cater well for that, apparently. Nonetheless, requests defined in the OpenAPI spec will be registered and their i/o data validated against that spec - all automatically.

The trade-off to this ad-hoc routing solution is that _only_ OpenAPI routes can exist in this application. Given that it's a REST API, I doubt this will prove insufficient any time soon.