Add `CachingStream`
This now caches the emitted response body. This is necessary because the original stream is one-time-use, as it wraps a non-rewindable iterator (containing ultimately DB records). This meant that, if the body (JSON) had to be validated before emitting (OpenAPI), it could not be emitted as the whole stream would be read, and could not be rewound.

The caching stream fixes that.